PR指摘したこと修正、Warning削減

+++ NinjaResidence/BaseGimmick.h
@@ -29,8 +29,8 @@ protected:
 
 	float m_QuantityOfMovement = 0.f;
 
-	float m_GimmickPosX;
-	float m_GimmickPosY;
+	int m_GimmickPosX;
+	int m_GimmickPosY;
 
 	float m_WorldPosLeft;
 	float m_WorldPosRight;
